Create Day1.cpp
Advent of Code 2023 Day 1 code. This code takes the first digit from a string, and combines it with the last digit of a string to create a two digit number. This is done for 1000 strings in a separate file and the total amount of all of these strings is submitted.